requirements.txt
snowsql
xlrd
snowflake-connector-python
spark
pandas
pyspark
airflow
jupyter notebook
papermill

